requirements.txt
PyDrive==1.3.1
requests==2.25.1
dropbox==11.12.0
picamera==1.13
piexif==1.1.3
Pillow==8.2.0
pyowm==3.2.0
pysftp==0.2.9
python_crontab==2.5.1
python_dateutil==2.8.1
twython==3.8.2
